# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "AAIF",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.30.0"
  },
  "paths": {
...
    "/dictation/models/{model_id}/download": {
      "get": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "get_download_progress",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadProgress"
                }
              }
            }
          },
          "404": {
            "description": "Download not found"
          }
        }
      },
      "post": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "download_model",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Download started"
          },
          "400": {
            "description": "Download already in progress"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "super::routes::dictation"
        ],
        "operationId": "cancel_download",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled"
          },
          "404": {
            "description": "Download not found"
          }
        }
      }
    },
...
    "/local-inference/download": {
      "post": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "download_hf_model",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Download started",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/local-inference/models/{model_id}": {
      "delete": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "delete_local_model",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model deleted"
          },
          "404": {
            "description": "Model not found"
          }
        }
      }
    },
    "/local-inference/models/{model_id}/download": {
      "get": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "get_local_model_download_progress",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadProgress"
                }
              }
            }
          },
          "404": {
            "description": "No active download"
          }
        }
      },
      "delete": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "cancel_local_model_download",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled"
          },
          "404": {
            "description": "No active download"
          }
        }
      }
    },
    "/local-inference/models/{model_id}/settings": {
      "get": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "get_model_settings",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettings"
                }
              }
            }
          },
          "404": {
            "description": "Model not found"
          }
        }
      },
      "put": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "update_model_settings",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettings"
                }
              }
            }
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Failed to save settings"
          }
        }
      }
    },
    "/local-inference/repo/{author}/{repo}/files": {
      "get": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "get_repo_files",
        "parameters": [
          {
            "name": "author",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GGUF files in the repo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoVariantsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/local-inference/search": {
      "get": {
        "tags": [
          "super::routes::local_inference"
        ],
        "operationId": "search_hf_models",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HfModelInfo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Search failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadModelRequest": {
        "type": "object",
        "required": [
          "spec"
        ],
        "properties": {
          "spec": {
            "type": "string",
            "description": "Model spec like \"bartowski/Llama-3.2-3B-Instruct-GGUF:Q4_K_M\""
          }
        }
      },
      "DownloadProgress": {
        "type": "object",
        "required": [
          "model_id",
          "status",
          "bytes_downloaded",
          "total_bytes",
          "progress_percent"
        ],
        "properties": {
          "bytes_downloaded": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes downloaded so far",
            "minimum": 0
          },
          "error": {
            "type": "string",
            "description": "Error message if failed",
            "nullable": true
          },
          "eta_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Estimated time remaining in seconds",
            "nullable": true,
            "minimum": 0
          },
          "model_id": {
            "type": "string",
            "description": "Model ID being downloaded"
          },
          "progress_percent": {
            "type": "number",
            "format": "float",
            "description": "Download progress percentage (0-100)"
          },
          "speed_bps": {
            "type": "integer",
            "format": "int64",
            "description": "Download speed in bytes per second",
            "nullable": true,
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/DownloadStatus"
          },
          "total_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Total bytes to download",
            "minimum": 0
          }
        }
      },
      "DownloadStatus": {
        "type": "string",
        "enum": [
          "downloading",
          "completed",
          "failed",
          "cancelled"
        ]
      },
...
      "HfGgufFile": {
        "type": "object",
        "description": "A single downloadable GGUF file (used internally and for downloads).",
        "required": [
          "filename",
          "size_bytes",
          "quantization",
          "download_url"
        ],
        "properties": {
          "download_url": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "quantization": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "HfModelInfo": {
        "type": "object",
        "required": [
          "repo_id",
          "author",
          "model_name",
          "downloads",
          "gguf_files"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "downloads": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "gguf_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HfGgufFile"
            }
          },
          "model_name": {
            "type": "string"
          },
          "repo_id": {
            "type": "string"
          }
        }
      },
      "HfQuantVariant": {
        "type": "object",
        "description": "A quantization variant — groups sharded files into one logical entry.",
        "required": [
          "quantization",
          "size_bytes",
          "filename",
          "download_url",
          "description",
          "quality_rank"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "download_url": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "quality_rank": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "quantization": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ModelSettings": {
        "type": "object",
        "properties": {
          "context_size": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "enable_thinking": {
            "type": "boolean"
          },
          "flash_attention": {
            "type": "boolean",
            "nullable": true
          },
          "frequency_penalty": {
            "type": "number",
            "format": "float"
          },
          "max_output_tokens": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "n_batch": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "n_gpu_layers": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "n_threads": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "native_tool_calling": {
            "type": "boolean"
          },
          "presence_penalty": {
            "type": "number",
            "format": "float"
          },
          "repeat_last_n": {
            "type": "integer",
            "format": "int32"
          },
          "repeat_penalty": {
            "type": "number",
            "format": "float"
          },
          "sampling": {
            "$ref": "#/components/schemas/SamplingConfig"
          },
          "use_jinja": {
            "type": "boolean"
          },
          "use_mlock": {
            "type": "boolean"
          }
        }
      },
...
      "RepoVariantsResponse": {
        "type": "object",
        "required": [
          "variants",
          "available_memory_bytes",
          "downloaded_quants"
        ],
        "properties": {
          "available_memory_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "downloaded_quants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recommended_index": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HfQuantVariant"
            }
          }
        }
      },
...
      "SamplingConfig": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Greedy"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "temperature",
              "top_k",
              "top_p",
              "min_p",
              "type"
            ],
            "properties": {
              "min_p": {
                "type": "number",
                "format": "float"
              },
              "seed": {
                "type": "integer",
                "format": "int32",
                "nullable": true,
                "minimum": 0
              },
              "temperature": {
                "type": "number",
                "format": "float"
              },
              "top_k": {
                "type": "integer",
                "format": "int32"
              },
              "top_p": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Temperature"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "tau",
              "eta",
              "type"
            ],
            "properties": {
              "eta": {
                "type": "number",
                "format": "float"
              },
              "seed": {
                "type": "integer",
                "format": "int32",
                "nullable": true,
                "minimum": 0
              },
              "tau": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "MirostatV2"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
    }
  }
}